refactor(webhookconfigurations): remove asset method and split tests
