+++ pkg/gatherers/clusterconfig/mutatingwebhookconfigurations_test.go
@@ -8,32 +8,7 @@ import (
 	admissionregistrationv1 "k8s.io/api/admissionregistration/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	kubefake "k8s.io/client-go/kubernetes/fake"
-
-	"github.com/openshift/insights-operator/pkg/record"
 )
-
-func Test_gatherValidatingWebhookConfigurations(t *testing.T) {
-	client := kubefake.NewSimpleClientset().AdmissionregistrationV1()
-	_, err := client.ValidatingWebhookConfigurations().Create(context.TODO(), &admissionregistrationv1.ValidatingWebhookConfiguration{
-		ObjectMeta: metav1.ObjectMeta{
-			Name: "webhook_config",
-		},
-		Webhooks: []admissionregistrationv1.ValidatingWebhook{
-			{
-				Name: "webhook1",
-			},
-			{
-				Name: "webhook2",
-			},
-		},
-	}, metav1.CreateOptions{})
-	assert.NoError(t, err)
-
-	records, errs := gatherValidatingWebhookConfigurations(context.TODO(), client)
-	assert.Empty(t, errs)
-
-	assertWebhookConfigurations(t, records, "config/validatingwebhookconfigurations/webhook_config")
-}
 
 func Test_gatherMutatingWebhookConfigurations(t *testing.T) {
 	client := kubefake.NewSimpleClientset().AdmissionregistrationV1()
@@ -55,12 +30,8 @@ func Test_gatherMutatingWebhookConfigurations(t *testing.T) {
 	records, errs := gatherMutatingWebhookConfigurations(context.TODO(), client)
 	assert.Empty(t, errs)
 
-	assertWebhookConfigurations(t, records, "config/mutatingwebhookconfigurations/webhook_config")
-}
-
-func assertWebhookConfigurations(t *testing.T, records []record.Record, expectedName string) {
 	assert.Len(t, records, 1)
-	assert.Equal(t, records[0].Name, expectedName)
+	assert.Equal(t, records[0].Name, "config/mutatingwebhookconfigurations/webhook_config")
 
 	configurationBytes, err := records[0].Item.Marshal()
 	assert.NoError(t, err)
